split module into components
