+++ hs-abci-examples/nameservice/Setup.hs
