+++ hs-abci-examples/nameservice/src/Nameservice/Modules/Nameservice/Messages.hs
@@ -1,0 +1,22 @@
+module Nameservice.Modules.Nameservice.Messages where
+
+import           Nameservice.Modules.Nameservice.Types (Name)
+import           Nameservice.Modules.Token             (Address, Amount)
+
+data MsgSetName =  MsgSetName
+  { msgSetNameName  :: Name
+  , msgSetNameValue :: String
+  , msgSetNameOwner :: Address
+  }
+
+data MsgDeletename = MsgDeleteName
+  { msgDeleteNameName  :: Name
+  , msgDeleteNameOwner :: Address
+  }
+
+data MsgBuyName = MsgBuyName
+    { msgBuyNameName  :: Name
+    , msgBuyNameValue :: String
+    , msgBuyNameBuyer :: Address
+    , msgBuyNameBid   :: Amount
+    }
